+++ src/main/java/org/mbari/mxm/db/parameter/ParameterDao.java
@@ -56,4 +56,16 @@ public interface ParameterDao {
   )
   @GetGeneratedKeys
   Parameter insertParameter(@BindBean Parameter pl);
+
+  @SqlUpdate(
+    """
+      delete from parameters
+      where provider_id    = :providerId
+        and mission_tpl_id = :missionTplId
+      """
+  )
+  Integer deleteForMissionTemplate(@Bind("providerId") String providerId,
+                                   @Bind("missionTplId") String missionTplId
+  );
+
 }